Module tauri_utils::acl

Access Control List types.

Stability

This is a core functionality that is not considered part of the stable API. If you use it, note that it may include breaking changes in the future.

These items are intended to be non-breaking from a de/serialization standpoint only. Using and modifying existing config values will try to avoid breaking changes, but they are free to add fields in the future - causing breaking changes for creating and full destructuring.

To avoid this, ignore unknown fields when destructuring with the {my, config, ..} pattern. If you need to create the Rust config directly without deserializing, then create the struct the Struct Update Syntax with ..Default::default(), which may need a #[allow(clippy::needless_update)] attribute if you are declaring all fields.

Re-exports

pub use self::identifier::*;

pub use self::value::*;

Modules

capability

End-user abstraction for selecting permissions a window has access to.

identifier

Identifier for plugins.

manifest

Plugin ACL types.

resolved

Resolved ACL for runtime usage.

value

A Value that is used instead of toml::Value or serde_json::Value to support both formats.

Structs

Commands

Allowed and denied commands inside a permission.

Permission

Descriptions of explicit privileges of commands.

PermissionSet

A set of direct permissions grouped together under a new name.

RemoteUrlPattern

UrlPattern for ExecutionContext::Remote.

Scopes

An argument for fine grained behavior control of Tauri commands.

Enums

Error

Possible errors while processing ACL files.

ExecutionContext

Execution context of an IPC call.

Constants

ACL_MANIFESTS_FILE_NAME

Known acl manifests file

APP_ACL_KEY

Known ACL key for the app permissions.

CAPABILITIES_FILE_NAME

Known capabilityies file

PERMISSION_SCHEMAS_FOLDER_NAME

Known foldername of the permission schema files

PERMISSION_SCHEMA_FILE_NAME

Known filename of the permission schema JSON file