Module tauri_utils::acl::capability

End-user abstraction for selecting permissions a window has access to.

Structs

Capability

A grouping and boundary mechanism developers can use to isolate access to the IPC layer.

CapabilityRemote

Configuration for remote URLs that are associated with the capability.

Enums

CapabilityFile

Capability formats accepted in a capability file.

PermissionEntry

An entry for a permission value in a Capability can be either a raw permission Identifier or an object that references a permission and extends its scope.