Module tauri_utils::config::parse

Items to help with parsing content into a Config.

Enums

ConfigError

Represents all the errors that can happen while reading the config.

Constants

EXTENSIONS_ENABLED

All extensions that are currently enabled.

EXTENSIONS_SUPPORTED

All extensions that are possibly supported, but perhaps not enabled.

Functions

does_supported_extension_exist

Check if a supported config file exists at path.

get_platform_config_filename

Gets the platform configuration file name.

parse

Parse the config from path, including alternative formats.

parse_json

“Low-level” helper to parse JSON into a Config.

parse_json_value

“Low-level” helper to parse JSON into a JSON Value.

parse_value

See parse for specifics, returns a JSON Value instead of Config.

read_from

Reads the configuration from the given root directory.

read_platform

Reads the platform-specific configuration file from the given root directory if it exists.