Struct tauri_utils::config::NsisConfig

pub struct NsisConfig {
    pub template: Option<PathBuf>,
    pub header_image: Option<PathBuf>,
    pub sidebar_image: Option<PathBuf>,
    pub installer_icon: Option<PathBuf>,
    pub install_mode: NSISInstallerMode,
    pub languages: Option<Vec<String>>,
    pub custom_language_files: Option<HashMap<String, PathBuf>>,
    pub display_language_selector: bool,
    pub compression: NsisCompression,
    pub start_menu_folder: Option<String>,
    pub installer_hooks: Option<PathBuf>,
    pub minimum_webview2_version: Option<String>,
}

Configuration for the Installer bundle using NSIS.

Fields
template: Option<PathBuf>

A custom .nsi template to use.

header_image: Option<PathBuf>

The path to a bitmap file to display on the header of installers pages.

The recommended dimensions are 150px x 57px.

sidebar_image: Option<PathBuf>

The path to a bitmap file for the Welcome page and the Finish page.

The recommended dimensions are 164px x 314px.

installer_icon: Option<PathBuf>

The path to an icon file used as the installer icon.

install_mode: NSISInstallerMode

Whether the installation will be for all users or just the current user.

languages: Option<Vec<String>>

A list of installer languages. By default the OS language is used. If the OS language is not in the list of languages, the first language will be used. To allow the user to select the language, set display_language_selector to true.

See https://github.com/kichik/nsis/tree/9465c08046f00ccb6eda985abbdbf52c275c6c4d/Contrib/Language%20files for the complete list of languages.

custom_language_files: Option<HashMap<String, PathBuf>>

A key-value pair where the key is the language and the value is the path to a custom .nsh file that holds the translated text for tauriโs custom messages.

See https://github.com/tauri-apps/tauri/blob/dev/crates/tauri-bundler/src/bundle/windows/nsis/languages/English.nsh for an example .nsh file.

Note: the key must be a valid NSIS language and it must be added to NsisConfig languages array,

display_language_selector: bool

Whether to display a language selector dialog before the installer and uninstaller windows are rendered or not. By default the OS language is selected, with a fallback to the first language in the languages array.

compression: NsisCompression

Set the compression algorithm used to compress files in the installer.

See https://nsis.sourceforge.io/Reference/SetCompressor

start_menu_folder: Option<String>

Set the folder name for the start menu shortcut.

Use this option if you have multiple apps and wish to group their shortcuts under one folder or if you generally prefer to set your shortcut inside a folder.

Examples:

AwesomePublisher, shortcut will be placed in %AppData%\Microsoft\Windows\Start Menu\Programs\AwesomePublisher\<your-app>.lnk

If unset, shortcut will be placed in %AppData%\Microsoft\Windows\Start Menu\Programs\<your-app>.lnk

installer_hooks: Option<PathBuf>

A path to a .nsh file that contains special NSIS macros to be hooked into the main installer.nsi script.

Supported hooks are:

NSIS_HOOK_PREINSTALL: This hook runs before copying files, setting registry key values and creating shortcuts.

NSIS_HOOK_POSTINSTALL: This hook runs after the installer has finished copying all files, setting the registry keys and created shortcuts.

NSIS_HOOK_PREUNINSTALL: This hook runs before removing any files, registry keys and shortcuts.

NSIS_HOOK_POSTUNINSTALL: This hook runs after files, registry keys and shortcuts have been removed.

Example

!macro NSIS_HOOK_PREINSTALL
  MessageBox MB_OK "PreInstall"
!macroend

!macro NSIS_HOOK_POSTINSTALL
  MessageBox MB_OK "PostInstall"
!macroend

!macro NSIS_HOOK_PREUNINSTALL
  MessageBox MB_OK "PreUnInstall"
!macroend

!macro NSIS_HOOK_POSTUNINSTALL
  MessageBox MB_OK "PostUninstall"
!macroend

minimum_webview2_version: Option<String>

Try to ensure that the WebView2 version is equal to or newer than this version, if the userโs WebView2 is older than this version, the installer will try to trigger a WebView2 update.