Trait tauri_utils::config::HeaderAddition

pub trait HeaderAddition {
    // Required method
    fn add_configured_headers(self, headers: Option<&HeaderConfig>) -> Builder;
}

A trait which implements on the Builder of the http create

Must add headers defined in the tauri configuration file to http responses

Required Methods

fn add_configured_headers(self, headers: Option<&HeaderConfig>) -> Builder

adds all headers defined on the config file, given the current HeaderConfig