Module tauri_utils::tokens

Utilities to implement ToTokens for a type.

Functions

json_value_lit

Create a serde_json::Value constructor TokenStream

json_value_number_lit

Create a serde_json::Value variant TokenStream for a number

map_lit

Create a map constructor, mapping keys and values with other TokenStreams.

opt_lit

Create an Option constructor TokenStream.

opt_lit_owned

Create an Option constructor TokenStream over an owned ToTokens impl type.

opt_str_lit

Helper function to combine an opt_lit with str_lit.

opt_vec_lit

Helper function to combine an opt_lit with a list of str_lit

path_buf_lit

Create a PathBuf constructor TokenStream.

str_lit

Create a String constructor TokenStream.

url_lit

Creates a Url constructor TokenStream.

vec_lit

Create a Vec constructor, mapping items with a function that spits out TokenStreams.