Crate tauri_utils

Tauri utility helpers

Modules

assets

The Assets module allows you to read files that have been bundled by tauri during both compile time and runtime.

config

The Tauri configuration used at runtime.

html

The module to process HTML in Tauri.

io

IO helpers.

mime_type

Determine a mime type from a URI or file contents.

pattern

Application pattern.

platform

Platform helper functions.

resources

Prepare application resources and sidecars.

Macros

consume_unused_variable

Suppresses the unused-variable warnings of the given inputs.

debug_eprintln

Prints to the standard error, with a newline.

Structs

Env

Information about environment variables.

PackageInfo

tauri::App package information.

Enums

Error

The error type of tauri-utils.

Theme

System theme.

TitleBarStyle

How the window title bar should be displayed on macOS.

Type Definitions

Result

The result type of tauri-utils.