Module tauri_utils::config

The Tauri configuration used at runtime.

It is pulled from a tauri.conf.json file and the Config struct is generated at compile time.

Stability

This is a core functionality that is not considered part of the stable API. If you use it, note that it may include breaking changes in the future.

These items are intended to be non-breaking from a de/serialization standpoint only. Using and modifying existing config values will try to avoid breaking changes, but they are free to add fields in the future - causing breaking changes for creating and full destructuring.

To avoid this, ignore unknown fields when destructuring with the {my, config, ..} pattern. If you need to create the Rust config directly without deserializing, then create the struct the Struct Update Syntax with ..Default::default(), which may need a #[allow(clippy::needless_update)] attribute if you are declaring all fields.

Re-exports

pub use self::parse::parse;

Modules

parse

Items to help with parsing content into a Config.

Structs

AndroidConfig

General configuration for the iOS target.

AppConfig

The App configuration object.

AppImageConfig

Configuration for AppImage bundles.

AssetProtocolConfig

Config for the asset custom protocol.

AssociationExt

An extension for a FileAssociation.

BuildConfig

The Build configuration object.

BundleConfig

Configuration for tauri-bundler.

Color

a tuple struct of RGBA colors. Each value has minimum of 0 and maximum of 255.

Config

The Tauri configuration object. It is read from a file where you can define your frontend assets, configure the bundler and define a tray icon.

DebConfig

Configuration for Debian (.deb) bundles.

DeepLinkProtocol

Deep link protocol configuration.

DmgConfig

Configuration for Apple Disk Image (.dmg) bundles.

FileAssociation

File association

IosConfig

General configuration for the iOS target.

LinuxConfig

Configuration for Linux bundles.

MacConfig

Configuration for the macOS bundles.

NsisConfig

Configuration for the Installer bundle using NSIS.

PluginConfig

The plugin configs holds a HashMap mapping a plugin name to its configuration object.

Position

Position coordinates struct.

RpmConfig

Configuration for RPM bundles.

SecurityConfig

Security configuration.

Size

Size of the window.

TrayIconConfig

Configuration for application tray icon.

WindowConfig

The window configuration object.

WindowEffectsConfig

The window effects configuration object

WindowsConfig

Windows bundler configuration.

WixConfig

Configuration for the MSI bundle using WiX.

WixLanguageConfig

Configuration for a target language for the WiX build.

Enums

BeforeDevCommand

Describes the shell command to run before tauri dev.

BundleResources

Definition for bundle resources. Can be either a list of paths to include or a map of source to target paths.

BundleTarget

Targets to bundle. Each value is case insensitive.

BundleType

A bundle referenced by tauri-bundler.

BundleTypeRole

macOS-only. Corresponds to CFBundleTypeRole

CapabilityEntry

A capability entry which can be either an inlined capability or a reference to a capability defined on its own file.

Csp

A Content-Security-Policy definition. See https://developer.mozilla.org/en-US/docs/Web/HTTP/CSP.

CspDirectiveSources

A Content-Security-Policy directive source list. See https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/Sources#sources.

CustomSignCommandConfig

Custom Signing Command configuration.

DisabledCspModificationKind

The possible values for the dangerous_disable_asset_csp_modification config option.

FrontendDist

Defines the URL or assets to embed in the application.

FsScope

Protocol scope definition. It is a list of glob patterns that restrict the API access from the webview.

HookCommand

Describes a shell command to be executed when a CLI hook is triggered.

NSISInstallerMode

Install Modes for the NSIS installer.

NsisCompression

Compression algorithms used in the NSIS installer.

PatternKind

The application pattern.

Updater

Updater type

V1Compatible

Generates lagacy zipped v1 compatible updaters

WebviewInstallMode

Install modes for the Webview2 runtime. Note that for the updater bundle Self::DownloadBootstrapper is used.

WebviewUrl

An URL to open on a Tauri webview window.

WixLanguage

The languages to build using WiX.