Module tauri_utils::config

The Tauri configuration used at runtime.

It is pulled from a tauri.conf.json file and the Config struct is generated at compile time.

Stability

This is a core functionality that is not considered part of the stable API. If you use it, note that it may include breaking changes in the future.

Re-exports

pub use self::parse::parse;

Modules

parse

Items to help with parsing content into a Config.

Structs

AllowlistConfig

Allowlist configuration. The allowlist is a translation of the Cargo allowlist features.

AppAllowlistConfig

Allowlist for the app APIs.

AppImageConfig

Configuration for AppImage bundles.

BuildConfig

The Build configuration object.

BundleConfig

Configuration for tauri-bundler.

CliArg

A CLI argument definition.

CliConfig

describes a CLI configuration

ClipboardAllowlistConfig

Allowlist for the clipboard APIs.

Config

The Tauri configuration object. It is read from a file where you can define your frontend assets, configure the bundler, enable the app updater, define a system tray, enable APIs via the allowlist and more.

DebConfig

Configuration for Debian (.deb) bundles.

DialogAllowlistConfig

Allowlist for the dialog APIs.

FsAllowlistConfig

Allowlist for the file system APIs.

GlobalShortcutAllowlistConfig

Allowlist for the global shortcut APIs.

HttpAllowlistConfig

Allowlist for the HTTP APIs.

HttpAllowlistScope

HTTP API scope definition. It is a list of URLs that can be accessed by the webview when using the HTTP APIs. The scoped URL is matched against the request URL using a glob pattern.

MacConfig

Configuration for the macOS bundles.

NotificationAllowlistConfig

Allowlist for the notification APIs.

NsisConfig

Configuration for the Installer bundle using NSIS.

OsAllowlistConfig

Allowlist for the OS APIs.

PackageConfig

The package configuration.

PathAllowlistConfig

Allowlist for the path APIs.

PluginConfig

The plugin configs holds a HashMap mapping a plugin name to its configuration object.

ProcessAllowlistConfig

Allowlist for the process APIs.

ProtocolAllowlistConfig

Allowlist for the custom protocols.

RemoteDomainAccessScope

External command access definition.

SecurityConfig

Security configuration.

ShellAllowedCommand

A command allowed to be executed by the webview API.

ShellAllowlistConfig

Allowlist for the shell APIs.

ShellAllowlistScope

Shell scope definition. It is a list of command names and associated CLI arguments that restrict the API access from the webview.

SystemTrayConfig

Configuration for application system tray icon.

TauriConfig

The Tauri configuration object.

UpdaterConfig

The Updater configuration object.

UpdaterEndpoint

A URL to an updater server.

UpdaterWindowsConfig

The updater configuration for Windows.

WindowAllowlistConfig

Allowlist for the window APIs.

WindowConfig

The window configuration object.

WindowsConfig

Windows bundler configuration.

WixConfig

Configuration for the MSI bundle using WiX.

WixLanguageConfig

Configuration for a target language for the WiX build.

Enums

AppUrl

Defines the URL or assets to embed in the application.

BeforeDevCommand

Describes the shell command to run before tauri dev.

BundleTarget

Targets to bundle. Each value is case insensitive.

BundleType

A bundle referenced by tauri-bundler.

Csp

A Content-Security-Policy definition. See https://developer.mozilla.org/en-US/docs/Web/HTTP/CSP.

CspDirectiveSources

A Content-Security-Policy directive source list. See https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/Sources#sources.

DisabledCspModificationKind

The possible values for the dangerous_disable_asset_csp_modification config option.

FsAllowlistScope

Filesystem scope definition. It is a list of glob patterns that restrict the API access from the webview.

HookCommand

Describes a shell command to be executed when a CLI hook is triggered.

NSISInstallerMode

Install Modes for the NSIS installer.

PatternKind

The application pattern.

ShellAllowedArg

A command argument allowed to be executed by the webview API.

ShellAllowedArgs

A set of command arguments allowed to be executed by the webview API.

ShellAllowlistOpen

Defines the shell > open api scope.

WebviewInstallMode

Install modes for the Webview2 runtime. Note that for the updater bundle Self::DownloadBootstrapper is used.

WindowUrl

An URL to open on a Tauri webview window.

WindowsUpdateInstallMode

Install modes for the Windows update.

WixLanguage

The languages to build using WiX.

Traits

Allowlist

Defines an allowlist type.