Module tauri_utils::config::parse

Items to help with parsing content into a Config.

Enums

ConfigError

Represents all the errors that can happen while reading the config.

ConfigFormat

The available configuration formats.

Constants

ENABLED_FORMATS

All configuration formats that are currently enabled.

EXTENSIONS_SUPPORTED

All extensions that are possibly supported, but perhaps not enabled.

SUPPORTED_FORMATS

All configuration formats that are possibly supported, but perhaps not enabled.

Functions

does_supported_file_name_exist

Check if a supported config file exists at path.

folder_has_configuration_file

Determines if the given folder has a configuration file.

is_configuration_file

Determines if the given file path represents a Tauri configuration file.

parse

Parse the config from path, including alternative formats.

parse_json

“Low-level” helper to parse JSON into a Config.

parse_json_value

“Low-level” helper to parse JSON into a JSON Value.

parse_value

See parse for specifics, returns a JSON Value instead of Config.

read_from

Reads the configuration from the given root directory.

read_platform

Reads the platform-specific configuration file from the given root directory if it exists.