Struct tauri_utils::config::WindowAllowlistConfig

pub struct WindowAllowlistConfig {

    pub all: bool,
    pub create: bool,
    pub center: bool,
    pub request_user_attention: bool,
    pub set_resizable: bool,
    pub set_maximizable: bool,
    pub set_minimizable: bool,
    pub set_closable: bool,
    pub set_title: bool,
    pub maximize: bool,
    pub unmaximize: bool,
    pub minimize: bool,
    pub unminimize: bool,
    pub show: bool,
    pub hide: bool,
    pub close: bool,
    pub set_decorations: bool,
    pub set_always_on_top: bool,
    pub set_content_protected: bool,
    pub set_size: bool,
    pub set_min_size: bool,
    pub set_max_size: bool,
    pub set_position: bool,
    pub set_fullscreen: bool,
    pub set_focus: bool,
    pub set_icon: bool,
    pub set_skip_taskbar: bool,
    pub set_cursor_grab: bool,
    pub set_cursor_visible: bool,
    pub set_cursor_icon: bool,
    pub set_cursor_position: bool,
    pub set_ignore_cursor_events: bool,
    pub start_dragging: bool,
    pub print: bool,

}

Allowlist for the window APIs.

See more: https://tauri.app/v1/api/config#windowallowlistconfig

Fields
all: bool

Use this flag to enable all window API features.

create: bool

Allows dynamic window creation.

center: bool

Allows centering the window.

request_user_attention: bool

Allows requesting user attention on the window.

set_resizable: bool

Allows setting the resizable flag of the window.

set_maximizable: bool

Allows setting whether the window’s native maximize button is enabled or not.

set_minimizable: bool

Allows setting whether the window’s native minimize button is enabled or not.

set_closable: bool

Allows setting whether the window’s native close button is enabled or not.

set_title: bool

Allows changing the window title.

maximize: bool

Allows maximizing the window.

unmaximize: bool

Allows unmaximizing the window.

minimize: bool

Allows minimizing the window.

unminimize: bool

Allows unminimizing the window.

show: bool

Allows showing the window.

hide: bool

Allows hiding the window.

close: bool

Allows closing the window.

set_decorations: bool

Allows setting the decorations flag of the window.

set_always_on_top: bool

Allows setting the always_on_top flag of the window.

set_content_protected: bool

Allows preventing the window contents from being captured by other apps.

set_size: bool

Allows setting the window size.

set_min_size: bool

Allows setting the window minimum size.

set_max_size: bool

Allows setting the window maximum size.

set_position: bool

Allows changing the position of the window.

set_fullscreen: bool

Allows setting the fullscreen flag of the window.

set_focus: bool

Allows focusing the window.

set_icon: bool

Allows changing the window icon.

set_skip_taskbar: bool

Allows setting the skip_taskbar flag of the window.

set_cursor_grab: bool

Allows grabbing the cursor.

set_cursor_visible: bool

Allows setting the cursor visibility.

set_cursor_icon: bool

Allows changing the cursor icon.

set_cursor_position: bool

Allows setting the cursor position.

set_ignore_cursor_events: bool

Allows ignoring cursor events.

start_dragging: bool

Allows start dragging on the window.

print: bool

Allows opening the system dialog to print the window content.